added control derivatives to 6 dof
some control derivatives are really small or really really big. according to the paper all the control derivatives are atleast 3 times higher when compared to the experimental values, this is due to the linear assumption used in the paper. The value of b_k/b affects differently to different coeffs so hard to adjust all the values for now.